+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.3] - 2020-06-30
+
+### Fixed
+- Done event was not emitting to the output event.
+
+### Changed
+- File upload using xhr request is removed due to done event of upload output was not emitting.
+
 ## [1.2.2] - 2020-06-30
 
 ### Changed
@@ -73,6 +81,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Added File select directive
 - Uploading files in a single request
 
+[1.2.3]: https://github.com/jayprajapati857/ngx-uploader-directive/compare/1.2.2...1.2.3
 [1.2.2]: https://github.com/jayprajapati857/ngx-uploader-directive/compare/1.2.1...1.2.2
 [1.2.1]: https://github.com/jayprajapati857/ngx-uploader-directive/compare/1.2.0...1.2.1
 [1.2.0]: https://github.com/jayprajapati857/ngx-uploader-directive/compare/1.1.7...1.2.0